requirements.txt
fastapi
pydantic
beautifulsoup4
lxml
uvicorn
asyncpg
sqlmodel
httpx
aiosqlite
python-jose[jwt]
passlib[bcrypt]
email_validator
pydantic-settings
